% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIMisc.r
\name{pi_hist}
\alias{pi_hist}
\title{A Hist Function}
\usage{
pi_hist(x)
}
\arguments{
\item{x}{No default. A data.frame}
}
\description{
This function allows you to plot histograms for all columns in data.frame
}
\examples{
pi_hist()
}
\keyword{hist}
